Created a better parsing method for amassite metadata
Instead of comparing the first line exactly the first line is parsed via regex
and the resulting matches are then stripped of whitespace on the left and right
sides. The resulting value is then checked as metadata. Still case sensitive.

The functions inside of all tags are now also stripped of whitespace on the left
and right sides before being processed.

Fixes #3
Fixes #7

Signed-off-by: Asher Glick <asher.glick@gmail.com>